requirements.txt
fastapi
uvicorn
imageio[v3]
opencv-python
transformers
torch
Pillow
numpy
pandas
requests
